remove old unused static line area bar charts
