--- frontend/src/metabase/static-viz/components/LineAreaBarChart/index.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export { default } from "./LineAreaBarChart";
