--- frontend/src/metabase/static-viz/components/LineAreaBarChart/types.ts
@@ -1,11 +1,0 @@
-export type Colors = {
-  brand: string;
-  brandLight: string;
-  textLight: string;
-  textMedium: string;
-};
-
-export type Size = {
-  width: number;
-  height: number;
-};
