+++ frontend/src/metabase/static-viz/components/Values/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./values";
+export * from "./series";
