+++ frontend/src/metabase/static-viz/components/Values/utils/values.ts
