+++ frontend/src/metabase/static-viz/components/WaterfallChart/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./scales";
