--- frontend/src/metabase/static-viz/components/XYChart/GoalLine/index.ts
@@ -1,1 +1,0 @@
-export * from "./GoalLine";
