--- frontend/src/metabase/static-viz/components/XYChart/constants.ts
@@ -1,5 +1,0 @@
-export const MAX_ROTATED_TICK_WIDTH = 100;
-
-export const CHART_PADDING = 12;
-
-export const LABEL_PADDING = 12;
