--- frontend/src/metabase/static-viz/components/XYChart/index.ts
@@ -1,1 +1,0 @@
-export * from "./XYChart";
