--- frontend/src/metabase/static-viz/components/XYChart/utils/index.ts
@@ -1,6 +1,0 @@
-export * from "./scales";
-export * from "./series";
-export * from "./ticks";
-export * from "./margin";
-export * from "./bounds";
-export * from "./values";
